==> 01_BaseLayer_Consolidate.Rmd <==
---
title: "Download base data from BC data catalogue"
author: "Matt Coghill and Gen Perkins"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r packages, include=FALSE}

#devtools::install_github("ColinChisholm/pemgeneratr")
library("pemgeneratr")
library(tidyverse)
library(bcdata)
library(bcmaps)
library(sf)
library(stars)
library(raster)
library(stringr)
library(fasterize)
library(plyr)
library(foreach)

```

## Overview 

This script downloads the relevant spatial data for BEC zones, VRI, TEM, waterbodies and the road network used in stage 1 of PEM processing. Data is downloaded directly from the [BC Data Catalogue](https://catalogue.data.gov.bc.ca/dataset?download_audience=Public) using the [bcdata](https://github.com/bcgov/bcdata) package.

### Datasets to download 

* BEC - [Biogeoclimatic Ecosystems Classification](https://catalogue.data.gov.bc.ca/dataset/bec-map) are used to define and/or select specific "subzones" within a defined study area. Note this can be also pointed to different file where more updated version is available.

* Vegetation Resource Inventory (VRI) - This layer includes a variety of vegetation measure, including cutblock age, TEM data and ..... Detailed data standards can be found [here](https://www2.gov.bc.ca/gov/content/industry/forestry/managing-our-forest-resources/forest-inventory/data-management-and-access/vri-data-standards).  

* Freshwater Atlas -  The atlas is separated into the different types of waterbodies (lakes, rivers, wetlands, streams, man made, etc.), which requires a seperate download per type. Alternatively a single combined layer can be downloaded by is limited to linear data type. A parameter within the function can be set to "polygon" (the default option) or "linear".  

* Road network THIS STILL NEEEDS WORK 

* Road network - In previous works, the [raw road network](https://catalogue.data.gov.bc.ca/dataset/digital-road-atlas-dra-master-partially-attributed-roads) was found to be too detailed. The raw road network is filtered to only keep named roads, and then the up to date [Forest Service Road (FSR) layer](https://catalogue.data.gov.bc.ca/dataset/forest-tenure-road-segment-lines) is downloaded and merged with the filtered road network to produce a really good representation of where roads are actually located on the landscape.

* Fire - Fire and fire intensity are used used to assist in identifying areas where a high cost is applied to reduce sampling. This includes:  
-[current](https://catalogue.data.gov.bc.ca/dataset/fire-perimeters-current) 
-[historical](https://catalogue.data.gov.bc.ca/dataset/fire-perimeters-historical) polygons.
-[severity](https://catalogue.data.gov.bc.ca/dataset/c58a54e5-76b7-4921-94a7-b5998484e697). 

* Accessing the [consolidated cutblock layer](https://catalogue.data.gov.bc.ca/dataset/harvested-areas-of-bc-consolidated-cutblocks-) using the bcdata package. The FTen layer is also used to identify very recent cutblocks ()


#########################
# A number of datasets are not available through the bcdata package that require a manual clip: 

* Tenure (private ownership) 	WHSE_CADASTRE.PMBC_PARCEL_FABRIC_POLY_FA_SVW	BCGW
* Detailed roads layers. In some cases regionally specific roads are available through other collaborative projects (cummulative effects). These are not firectly available through the BCGW.


## Work-flow
This script relies on having a spatial file (AOI.gpkg) defining the study area (area of interest - AOI). Firstly the area of interest is defined along with input and output folders. 

```{r Parameters}

#AOI <- "Deception"
#AOI <- "KootInvCran"
AOI <- "KIC_SE"
#AOI <- "KIC_SW"
#AOI <- "KIC_NW"
#AOI <- "KIC_NE"

AOI_dir <- file.path(paste0("./",AOI, "_AOI"))
raw_dir <- file.path(AOI_dir, "0_raw_inputs")
derived_dir <- file.path(AOI_dir, "1_map_inputs")
cov_dir <- file.path(AOI_dir, "1_map_inputs", "covariates")
shape_dir <- file.path(AOI_dir, "0_raw_inputs", "base_layers")
dem_dir <- file.path(AOI_dir, "0_raw_inputs", "dem")
out_path <- shape_dir

source("./_functions/aoi_snap.R")

```

# Create template AOIs for the analysis

We will modify the aoi to ensure the area of interest can be easily stacked in future analysis. 

```{r}
epsg = 3005

aoi_raw <- st_read(file.path(shape_dir, "aoi.gpkg"), quiet = TRUE) %>%
  st_zm() %>%
  st_transform(3005)

e <- as(extent(aoi_raw), "SpatialPolygons") ## for use in map below.
aoi <- aoi_snap(aoi_raw, method = "expand")
aoi$id = 1  
st_write(aoi, file.path(shape_dir, "aoi_snapped.gpkg"))

```

Initially a single function is used to download BEC zone info, VRI shapes, TEM shapes (if applicable), water bodies, and roads all in one go. Each is clipped to the extent of the study area at each resolution, however with increasing size of study area this was converted to individual snippets/functions. 

```{r Function}

  in_aoi <- st_read(file.path(shape_dir, "aoi_snapped.gpkg"))

  bec_dir <- file.path("_data_allBC", "BECv12.gpkg")
  
  # checks that everything is set correctly 
  
   if(missing(in_aoi)) stop("'in_aoi' is missing with no default") 
    
   # Second, detect object type and convert where necessary
   if(!inherits(in_aoi, c("sf", "sfc")))
     stop("'in_aoi' is not an sf or sfc object.")
    
   if(is.null(out_path)) {
     stop("\rout_path is an invalid file path string")
   }
 
  # Detect the CRS of the sf object
  if(is.na(sf::st_crs(in_aoi)))
    stop("CRS is not assigned. Use sf::st_crs() to assign a valid CRS to in_aoi")

  if(sf::st_is_longlat(in_aoi)) {
    cat("Input CRS is Lat/Long format. Transforming to EPSG 3005 (BC Albers) for processing\n")
    epsg <- 3005L
    in_crs <- sf::st_crs(in_aoi)
    in_aoi <- sf::st_transform(in_aoi, epsg) %>% sf::st_set_agr("constant")
  } else {
    in_crs <- sf::st_crs(in_aoi)
    epsg <- in_crs$epsg
    in_aoi <- sf::st_set_agr(in_aoi, "constant")
    if(!is.numeric(epsg))
      stop("There was a problem retrieving the EPSG code from the in_aoi. Is it assigned properly?")
  }

  # Get largest bbox for downloading layers using the pemgeneratr function to match the raster inputs
  box <- sf::st_bbox(in_aoi)
  
  # Adjust max download size based on AOI
  options(bcdata.max_geom_pred_size = as.numeric(st_area(in_aoi)) + 10)
  
  # # Bec map - option 1) get Bec mapping from BCGW     
  # bec_sf <- bcmaps::bec(class = "sf", force = TRUE, ask = FALSE) %>% 
  #       sf::st_set_agr("constant")
  #     if(sf::st_crs(bec_sf) != sf::st_crs(in_aoi)) {
  #       box_trans <- sf::st_as_sfc(box) %>% 
  #         sf::st_buffer(10, nQuadSegs = 0) %>% 
  #         sf::st_transform(sf::st_crs(bec_sf)) %>% 
  #         sf::st_bbox()
  #       bec_sf <- sf::st_crop(bec_sf, box_trans) %>% 
  #         sf::st_transform(sf::st_crs(in_aoi)) %>% 
  #         sf::st_set_agr("constant")
  #     } 
  
  # alternatively get more updated bec from another source: 
  bec_sf <- st_read(bec_dir)  
  bec_sf <- st_cast(bec_sf, "MULTIPOLYGON")
  bec_sf_clip <- sf::st_crop(bec_sf, box) 
  bec_sf <- st_intersection(aoi_raw, bec_sf_clip) %>%
    dplyr::select(BGC_LABEL)
  
  st_write(bec_sf, file.path(out_path, "bec.gpkg"), delete_dsn = TRUE, 
           delete_layer = TRUE)
  
  # Download VRI
  vri <- bcdc_query_geodata("2ebb35d8-c82f-4a17-9c96-612ac3532d55", crs = epsg) %>% 
    bcdata::filter(INTERSECTS(aoi_raw)) %>% 
    bcdata::select(c("BCLCS_LEVEL_2","BCLCS_LEVEL_4","PROJ_AGE_CLASS_CD_1", "SPECIES_CD_1")) %>% # Treed sites
    collect() %>% 
    {if(nrow(.) > 0) st_intersection(., aoi_raw) else .}
  
  st_write(vri, file.path(out_path, "vri.gpkg"), delete_dsn = TRUE,
           delete_layer = TRUE)

  
# post process VRI data into classes: 
# Depending on the study area we want to focus on sampling in older areas - class 4 (60-80) or 5 (80 + )
# To do this we will define 2 vri age classess to exclude from the sampling area
  # class 1 and 2 (0 - 40 yrs)
  # class 1-3 (0 - 60 years)

#  vri <- st_read(file.path(out_path, "vri.gpkg")) %>%
#    dplyr::select(c("BCLCS_LEVEL_2","BCLCS_LEVEL_4","PROJ_AGE_CLASS_CD_1", "SPECIES_CD_1"))
  
  vri_class2 <- vri %>%
    dplyr::mutate(age_class = as.numeric(PROJ_AGE_CLASS_CD_1)) %>%
    dplyr::filter(age_class < 3)
  st_write(vri_class2, file.path(out_path, "vri_class1_2.gpkg"))
  
  vri_class3 <- vri %>%
    dplyr::mutate(age_class = as.numeric(PROJ_AGE_CLASS_CD_1)) %>%
    dplyr::filter(age_class == 3)
  st_write(vri_class3, file.path(out_path, "vri_class3.gpkg"))

  
######### this section is currently in test phase (related to removal fo deciduous areas in Date Creek after these were found to be highly selected and challanging to sample)
  
  # vri - deciduous leading - this needs to be appled for some 
  
  # STILL TO DO - for areas with deciduous leading (AT, EP) Aspen and paper burch these should be seperated using the code "SPECIES_CD_1" == AT|EP. 
  # ie important in Date Creek 
   
  vri_decid <- vri %>%
     dplyr::filter(SPECIES_CD_1 %in% c("AT","EP")) # note might need to adjust for some areas of interest
  
    st_write(vri_decid, file.path(out_path, "vri_decid.gpkg"))
   
######### end of test section   
   
  # Download recent cutblocks layer
  # Uses date filter which filters cutblock ages less than 20 years, or 7305 days
  cutblocks <- bcdc_query_geodata("b1b647a6-f271-42e0-9cd0-89ec24bce9f7", crs = epsg) %>% 
    bcdata::filter(INTERSECTS(aoi_raw)) %>% 
    collect() 
  
  cutblocks <- cutblocks %>% 
    {if(nrow(.) > 0) st_intersection(., in_aoi) else .} %>% 
  dplyr::filter(as.numeric(format(Sys.time(), "%Y")) - HARVEST_YEAR <= 20)

  st_write(cutblocks, file.path(out_path, "cutblocks.gpkg"), delete_dsn = TRUE,
           delete_layer = TRUE)
  
  # ften  - download latest harvest layer 
   ften <- bcdc_query_geodata("cff7b8f7-6897-444f-8c53-4bb93c7e9f8b", crs = epsg) %>% 
    bcdata::filter(INTERSECTS(aoi_raw)) %>% 
    bcdata::select("HARVEST_AUTH_STATUS_CODE", "ISSUE_DATE", "CURRENT_EXPIRY_DATE_CALC", "LIFE_CYCLE_STATUS_CODE", "FILE_STATUS_CODE") %>%# Treed sites
    collect() 
   
  ften <- ften %>%
    filter(ISSUE_DATE >2000)
  
  ften <- ften %>%
    dplyr::select("HARVEST_AUTH_STATUS_CODE", "ISSUE_DATE", "CURRENT_EXPIRY_DATE_CALC",
                  "LIFE_CYCLE_STATUS_CODE", "FILE_STATUS_CODE","FEATURE_AREA")
   
   st_write(ften, file.path(out_path, "ften.gpkg"), delete_dsn = TRUE,
            delete_layer = TRUE)
   
  
  # cutblocks.gpkg"))
  cutblocks <- st_read(file.path(out_path, "cutblocks.gpkg")) # cutblocks.gpkg"))
  ften <- st_read(file.path(out_path, "ften.gpkg")) # cutblocks.gpkg"))
  cutblocks_ften <- bind_rows(cutblocks, ften)
  cutblocks_ften <- bind_rows( ften, cutblocks)
  st_write(cutblocks_ften, file.path(out_path, "cutblocks_ften.gpkg"),delete_dsn = TRUE,
            delete_layer = TRUE)
  
  # TEM
  tem <- bcdc_query_geodata("0a83163b-a62f-4ce6-a9a1-21c228b0c0a3", crs = epsg) %>%
    bcdata::filter(INTERSECTS(in_aoi)) %>%
    collect() %>% 
    {if(nrow(.) > 0) st_intersection(., aoi_raw) else .}
  
  st_write(tem, file.path(out_path, "tem.gpkg"), delete_dsn = TRUE, 
           delete_layer = TRUE)
  
  # water layers, requires merging multiple layers
  
  # Use foreach in parallel to efficiently download multiple water layers
  water_records <- c("cb1e3aba-d3fe-4de1-a2d4-b8b6650fb1f6", # lakes
                     "f7dac054-efbf-402f-ab62-6fc4b32a619e", # rivers
                     "93b413d8-1840-4770-9629-641d74bd1cc6") # wetlands
  
  cl <- parallel::makeCluster(parallel::detectCores())
  doParallel::registerDoParallel(cl)
  
  waterbodies <- foreach(i = water_records, .combine = rbind, 
                         .packages = c("tidyverse", "bcdata", "sf")) %dopar%
      {
      bcdc_query_geodata(i, crs = epsg) %>% # lakes
        bcdata::filter(INTERSECTS(in_aoi)) %>% 
        collect() %>% 
        {if(nrow(.) > 0) st_intersection(., in_aoi) else NULL}
      }
  parallel::stopCluster(cl)
  
 
  # filter a subset of columns 
  waterbodies_sf <- waterbodies %>%
    dplyr::select(id, WATERBODY_TYPE, AREA_HA)
  
  st_write(waterbodies_sf, file.path(out_path, "water.gpkg"), delete_dsn = TRUE, 
           delete_layer = TRUE)
  
  # Download road network
  # The main road network layer has too many roads in it. Filter it down to only
  # include named roads and combine those with actual mapped FSR's
  message("\rDownlomapading Road network")
  roads <- bcdc_query_geodata("bb060417-b6e6-4548-b837-f9060d94743e", crs = epsg) %>% 
    bcdata::filter(
      BBOX(st_bbox(in_aoi), crs = paste0("EPSG:", epsg))) %>%#,
      #ROAD_NAME_ID > 0) %>% 
    collect() %>% 
    dplyr::select(id, ROAD_NAME_FULL, FEATURE_LENGTH_M) %>% 
    dplyr::rename(NAME = ROAD_NAME_FULL) %>% 
    {if(nrow(.) > 0) {
      st_intersection(., in_aoi) %>% 
        st_cast("MULTILINESTRING")
      } else .}
  
  fsr <- bcdc_query_geodata("9e5bfa62-2339-445e-bf67-81657180c682", crs = epsg) %>% 
    bcdata::filter(
      BBOX(st_bbox(in_aoi), crs = paste0("EPSG:", epsg)), 
      LIFE_CYCLE_STATUS_CODE == "ACTIVE") %>% 
    collect() %>% 
    dplyr::select(id, MAP_LABEL, FEATURE_LENGTH_M) %>% 
    dplyr::rename(NAME = MAP_LABEL) %>% 
    {if(nrow(.) > 0) {
      st_intersection(., in_aoi) %>% 
        st_cast("MULTILINESTRING")
    } else .}
  
  road_merge <- rbind(roads, fsr)
  
  st_write(road_merge, file.path(out_path, "road_network.gpkg"), delete_dsn = TRUE, 
           delete_layer = TRUE)
  
  
  roads <- st_read(file.path(out_path, "road_surroundings.gpkg"))
  names(roads)
 
  roads_vet <- st_read(file.path(out_path, "roads_vetted.gpkg"))
  
  # roads_vet <- st_read(file.path(out_path, "raw","SE_roads_to_add2.gpkg"))
  # #names(roads_vet)
  # roads_vet <- roads_vet %>% select(c(ROAD_SURFA, ROAD_CLASS)) %>% 
  #   mutate(ROAD_SURFA = "rough", 
  #          ROAD_CLASS = "resource")
  # 
  #   roads_vet <- bind_rows(roads, roads_vet)
  # st_write(roads_vet, file.path(out_path, "road_surroundingsV2.gpkg"))

  
  roads_vet <- bind_rows(roads, roads_vet)
  st_write(roads_vet, file.path(out_path, "road_vetted.gpkg"))
#   roads_vet <- st_read(file.path(out_path, "roads_vetted.gpkg"))
#  roadstoadd <- st_read(file.path(out_path, "raw", "roads_toaddd.gpkg"))  
  
#  roads <- bind_rows(roads, roadstoadd)
#  st_write(roads, file.path(out_path, "road_surroundings2.gpkg"))
  
  # download main towns (and wider cut of roads)
  # still needs work on output
  
  towns <- bcdc_query_geodata("b678c432-c5c1-4341-88db-0d6befa0c7f8",crs = epsg) %>%
    collect()
  
  st_write(towns, file.path(out_path, "major_towns_bc.gpkg"), delete_dsn = TRUE, 
           delete_layer = TRUE)
  
  # fire polygons
  fire_records <- c("cdfc2d7b-c046-4bf0-90ac-4897232619e1", 
                    "22c7cb44-1463-48f7-8e47-88857f207702")
    
  cl <- parallel::makeCluster(parallel::detectCores())
  doParallel::registerDoParallel(cl)
  
  fires <- foreach(i = fire_records, .combine = rbind, 
                         .packages = c("tidyverse", "bcdata", "sf")) %dopar% 
    {
      bcdc_query_geodata(i, crs = epsg) %>%
        bcdata::filter(INTERSECTS(in_aoi)) %>% 
        collect() %>% 
        {if(nrow(.) > 0) st_intersection(., in_aoi) %>% 
            dplyr::select(id, FIRE_NUMBER, VERSION_NUMBER, FIRE_YEAR, 
                          FIRE_SIZE_HECTARES, LOAD_DATE) %>% 
            dplyr::filter(as.numeric(format(Sys.time(), "%Y")) - FIRE_YEAR <= 20)}
    }
  
  parallel::stopCluster(cl)
  
  st_write(fires, file.path(out_path, "fire.gpkg"), delete_dsn = TRUE, 
           delete_layer = TRUE)
  
  # fire severity 
    fire_int <- bcdc_query_geodata("c58a54e5-76b7-4921-94a7-b5998484e697",crs = epsg) %>% 
    bcdata::filter(INTERSECTS(aoi_raw)) %>% 
    bcdata::select(c("FIRE_YEAR", "BURN_SEVERITY_RATING")) %>% # Treed sites
    collect()
    
  st_write(fire_int, file.path(out_path, "fire_int.gpkg"), delete_dsn = TRUE, 
           delete_layer = TRUE)

  # BC parks 
  #bcdc_search("parks and protected area")
  #bcdc_search("1130248f-f1a3-4956-8b2e-38d29d3e4af7")
  
    parks <- bcdc_query_geodata("1130248f-f1a3-4956-8b2e-38d29d3e4af7",crs = epsg) %>% 
    bcdata::filter(INTERSECTS(aoi_raw)) %>% 
    collect()
    
  st_write(parks, file.path(out_path, "parks.gpkg"), delete_dsn = TRUE, 
           delete_layer = TRUE)
  
  # National parks (if an option)
  national_parks <- bcdc_query_geodata("88e61a14-19a0-46ab-bdae-f68401d3d0fb",crs = epsg) %>% 
    bcdata::filter(INTERSECTS(aoi_raw)) %>% 
    collect()
  
   st_write(national_parks, file.path(out_path, "natparks.gpkg"), delete_dsn = TRUE, 
           delete_layer = TRUE)
   
  # transmission lines
  #bcdc_search("transmission")
   trans_line <-  bcdc_query_geodata("384d551b-dee1-4df8-8148-b3fcf865096a",crs = epsg) %>% 
    bcdata::filter(INTERSECTS(aoi_raw)) %>% 
    collect()
   
  st_write(trans_line, file.path(out_path, "translines.gpkg"), delete_dsn = TRUE, 
           delete_layer = TRUE)


  ## STILL A WORK IN PROGRESS
   
   # sedimentary layer. Currently testing this as part of KIC study area for an additional landscape scale analysis 
   
   
     # alternatively get more updated bec from another source: 
  geo_dir <- file.path("_data_allBC", "BC_digital_geology_gpkg","BC_digital_geology.gpkg")
  geo_sf <- st_read(geo_dir,  layer = "Bedrock_ll83_poly")
  geo_sf <- st_cast(geo_sf, "MULTIPOLYGON") %>%
    dplyr::select("upid","rock_class" )
  geo_sf <- sf::st_transform(geo_sf, sf::st_crs(in_aoi)) #%>% 
          #sf::st_set_agr("constant")
  geo_sf_clip <- sf::st_crop(geo_sf, box) 
  geo_sf <- st_intersection(aoi_raw, geo_sf_clip) %>%
    dplyr::select(rock_class, geom)
  
  st_write(geo_sf, file.path(out_path, "geo.gpkg"), delete_dsn = TRUE, 
           delete_layer = TRUE)

   
```

```{r}
# Post processing requirements

# 1) landownership

# As landownership needs to be downloaded separately we need to filter this for private ownership (remove anything except None or unknown)

#landownreship <- st_read(file.path(out_path, "raw", "dc_landownership.shp")) %>%
landownreship <- st_read(file.path(out_path, "raw", "parcels_se.gpkg")) %>%
#landownreship <- st_read(file.path(out_path, "raw", "wl_cadastre_parcel.shp")) %>%
  st_zm %>%
  dplyr::select(OWNER_TYPE)%>%
  dplyr::filter(OWNER_TYPE %in% c("Private", "Crown Provincial" , "Federal",
                           "Municipal" , "First Nations","Crown Agency",
                           "Mixed Ownership"))
 
st_write(landownreship, file.path(out_path, "private.gpkg"))
 

```


# Other data sets that might be useful but not essential for sample plans

```{r auxillary dataseyts}

# bc boundary
 bc_boundary <- bcdc_get_data('30aeb5c1-4285-46c8-b60b-15b1a6f4258b', resource = '3d72cf36-ab53-4a2a-9988-a883d7488384')
    
# st_write( bc_boundary, file.path(out_path, "fire_int.gpkg"), delete_dsn = TRUE, 
#           delete_layer = TRUE)

# Regional Distribt for BC 
try(
  bcdc_get_record("https://catalogue.data.gov.bc.ca/dataset/regional-districts-legally-defined-administrative-areas-of-bc")
)

rd <- bcdc_query_geodata('d1aff64e-dbfe-45a6-af97-582b7f6418b9',crs = 3005) %>%
    collect()
    
st_write(rd, file.path("_data_allBC", "regional_district.gpkg"), delete_dsn = TRUE, 
           delete_layer = TRUE)
  
  # Forest Distribt for BC 
try(
  bcdc_get_record("https://catalogue.data.gov.bc.ca/dataset/natural-resource-nr-district")
)

fdistict <- bcdc_query_geodata('0bc73892-e41f-41d0-8d8e-828c16139337',crs = 3005) %>%
    collect()
    
st_write(fdistict, file.path("_data_allBC", "forest_district.gpkg"), delete_dsn = TRUE, 
           delete_layer = TRUE)
  



  
```
