==> 04a_TrainingPt_TransectImport_QA.Rmd <==
---
title: "Import of Stage 1 Field Transect data for QA purposes"
subtitle: "by Gen Perkins"
date: "19/07/2022"
output: html_document
---

```{r global_options, include=FALSE }
require(knitr)

```

# read in librarys 

```{r setup, include=FALSE}
library(tidyverse)
library(sf)
library(stringr)
library(lwgeom)
library(foreach)
library(scales)
library(raster)
library(tidyverse)
library(raster)
library(fasterize)
library(tools)
library(dplyr)
library(data.table)
library(cowplot)
library(tidymodels)
library(themis)
library(ggplot2)
library(janitor)

```

## Introduction

This script contains functionality for the import data collected by Avenza following field collection of triangular paired transects for Stage 1 data collection.

1. Import transects from shapefiles or geopackage exported from Avenza PDF maps. 

2. Converts data to vector triangles (using point to point at 5m resolution). Specifically this converts the points to lines (or track log to lines), add the field sample data as attributes to the line. 

To run this script you will need:  
- Field data + tracks collected as .shp or .gpkg files. 
- Mapunit_Legend (csv file which will convert the codes to standard codes)

3. For each of the QA and contractor datasets, convert to standard xy locations using a 5m raster template. 

4. Compare QA and contractor data and calculate accuracy statistics







## 1. Set up folder structure for Stage 1 sampling 

```{r set up folder structure }

AOI <- "KootInvCran_AOI_QA"
aoi_short <- "KIC"
res <- 2.5 # set resolution
#AOI_dir <- file.path(".", paste0(AOI,"reference_data")) 
AOI_dir <- file.path(AOI) 
AOI_reference <- file.path(AOI_dir,"reference_data")
trans_layout <- file.path(AOI_reference , "transect_layout")
trans_data_QA <- file.path(AOI_dir, "QA_transect_data")
trans_data_cont <- file.path(AOI_dir, "contractor_transect_data")

map.key  <- read.csv(file.path(AOI_reference , "KIC_MapUnitLegend.csv"), 
                       stringsAsFactor = FALSE)

output_pnts_dir <- file.path(AOI_dir, "output")

raster_template  <- raster(file.path(AOI_reference, "template.tif"))
                  
if(!dir.exists(output_pnts_dir)) dir.create(output_pnts_dir, recursive = TRUE)

```

Setup functions 

```{r Functions}

source(here::here('_functions', 'extend_lines.R'))
source(here::here('_functions', 'make_lines.R'))
source(here::here('_functions', 'transect_sample.R'))
source(here::here('_functions', 'multiline_to_line.R'))
source(here::here('_functions', 'acc_metrix_QA.R'))

```


## 1) Import and clean Avenza field transect data

1. Import transects from shape files or geopackages exported from Avenza PDF maps. The shape files contain all data collected under the schema and so generally contain all transects collected to date on a single tablet. Need to separate using the original study design transect triangles.


```{r import base files to run analysis, tidy = TRUE, warning=FALSE}

trans <- list.files(trans_layout, pattern = ".gpkg$", full.names = TRUE, recursive = FALSE)

transect_layout <- foreach(x = trans, .combine = rbind) %do% {
  clhs_layers <- st_layers(x)
  lines <- which(clhs_layers[["geomtype"]] %in% c("Line String", "Multi Line String"))
  if(length(lines)) {
    transects <- foreach(y = clhs_layers$name[lines], .combine = rbind) %do% {
      transect <- st_read(x, y, quiet = TRUE) %>% 
        rename_all(recode, geom = "geometry") %>% 
        rename_all(.funs = tolower) %>%
        dplyr::select(id) %>% 
        mutate(id = as.character(id)) %>% 
        st_transform(3005)
    }
  } 
} 

transect_layout <- unique(transect_layout)

transect_layout_buf <- st_buffer(transect_layout, 10)

st_write(transect_layout, file.path(output_pnts_dir, "transect_layout_s1.gpkg"), delete_layer = TRUE)


```


2. Connect the placemarks waypoints into a traverse using the running number name assigned by PDF maps

The following chunk creates some basic folders and loads in data that will be used across all sampling types (a raster template and the original CLHS transects).


# QA DATA: 

```{r prepare raw data (unzip data and split into points and lines)}
## prepare point data sets for QA data set 

poc_column <- "x03_pt_type"
mapunit1_column <- "x04_mapunit"
mapunit2_column <- "x06_mapunit"
observer_column <- "x02_observe"
transition_column <- "x05_transit"
comment_column <- "x09_comment"


# note currently this is set up for shapefile inputs 

shp <- list.files(file.path(trans_data_QA), pattern = ".shp", full.names = TRUE, recursive = TRUE)

tracks <- list.files(file.path(trans_data_QA), pattern = "\\(1\\).shp", full.names = TRUE, recursive = TRUE)

points <- shp[!shp %in% tracks]
#points <- points[4]

all_points <- foreach(x = points, .combine = rbind) %do% {
  #x <- points[1]
  print(x)
      # when all the points are consolidated into a single geopackage 
     points_read <- st_read(x, quiet = TRUE) %>%
     st_transform(3005) %>% 
     st_zm() %>% 
     rename_all(.funs = tolower) %>%
     mutate(mapunit1 = UQ(rlang::sym(mapunit1_column)),
            mapunit2 = UQ(rlang::sym(mapunit2_column)),
            point_type  = UQ(rlang::sym(poc_column)),
            observer =  UQ(rlang::sym(observer_column)), 
            transition = UQ(rlang::sym(transition_column)),
            comments = UQ(rlang::sym(comment_column))) %>%
     st_join(., transect_layout_buf, join = st_intersects) %>%
     rename_all(.funs = tolower) %>%
     rename(transect_id = id)
  
     if("name" %in% names(points_read)){
       points_read <- points_read %>%
         mutate(order = as.numeric(gsub("Placemark ", "", name)))
     }
    points_read <- points_read %>%
     dplyr::select(order, mapunit1, mapunit2, point_type, transect_id, observer, transition, comments) %>%
     group_by(transect_id) %>%
     arrange(as.numeric(order), by_group = TRUE) %>%
     ungroup() 

} %>% 
  distinct(., .keep_all = TRUE) 

# catergorise data types:
points <- all_points %>%
  mutate(data_type = ifelse(is.na(transect_id), "incidental", "s1"))

# match to the baseunit key using mapunit legend & format observer column
points_final <- format_mapunit_names(points, map.key)
points_final <- fill_observer(points_final)
 
# covert to lines 
processed_transects <- make_lines(GPSPoints = points_final, 
                                  Transects = transect_layout_buf, 
                                  method = "pts2lines",  
                                  tBuffer = 20, PROJ = 3005) %>%
  mutate(mapunit1 = (trimws(mapunit1)),
         mapunit2 = (trimws(mapunit2))) %>%
  mutate(mapunit12 = paste0(mapunit1,"_", mapunit2)) %>%
  mutate(mapunit12 = gsub("_NA","", mapunit12)) %>%
  mutate(mapunit12 = gsub("_$","", mapunit12)) %>%
  dplyr::select(-TID, -ID)

st_write(processed_transects,  file.path(output_pnts_dir, "proc_s1_transects_QA.gpkg"), 
         delete_layer = TRUE)

st_write(points_final, file.path(output_pnts_dir, "s1_pts_QA.gpkg"), delete_layer = TRUE)

```

 ## CONTRACTOR DATA 

```{r}

# this might re quire prepocessing using the function "_standardise_avenza_input_field_names.R"

points <- list.files(file.path(trans_data_cont), pattern = ".gpkg", full.names = TRUE)

tracks <- list.files(file.path(trans_data_cont), pattern = "\\(1\\).gpkg", full.names = TRUE, recursive = TRUE)

points <- points[!points%in% tracks]
#points <- points[4]

all_points <- foreach(x = points, .combine = rbind) %do% {
  #x <- points[1]
  print(x)
      # when all the points are consolidated into a single geopackage 
     points_read <- st_read(x, quiet = TRUE) %>%
     st_transform(3005) %>% 
     st_zm() %>% 
     rename_all(.funs = tolower) %>%
     mutate(mapunit1 = UQ(rlang::sym(mapunit1_column)),
            mapunit2 = UQ(rlang::sym(mapunit2_column)),
            point_type  = UQ(rlang::sym(poc_column)),
            observer =  UQ(rlang::sym(observer_column)), 
            transition = UQ(rlang::sym(transition_column)),
            comments = UQ(rlang::sym(comment_column))) %>%
     st_join(., transect_layout_buf, join = st_intersects) %>%
     rename_all(.funs = tolower) %>%
     rename(transect_id = id)
  
     if("name" %in% names(points_read)){
       points_read <- points_read %>%
         mutate(order = as.numeric(gsub("Placemark ", "", name)))
     }
    points_read <- points_read %>%
     dplyr::select(order, mapunit1, mapunit2, point_type, transect_id, observer, transition, comments) %>%
     group_by(transect_id) %>%
     arrange(as.numeric(order), by_group = TRUE) %>%
     ungroup() 

} %>% 
  distinct(., .keep_all = TRUE) 

# catergorise data types:
points <- all_points %>%
  mutate(data_type = ifelse(is.na(transect_id), "incidental", "s1"))

# match to the baseunit key using mapunit legend & format observer column
points_final <- format_mapunit_names(points, map.key)
points_final <- fill_observer(points_final)
 
# covert to lines 
processed_transects <- make_lines(GPSPoints = points_final, 
                                  Transects = transect_layout_buf, 
                                  method = "pts2lines",  
                                  tBuffer = 20, PROJ = 3005) %>%
  mutate(mapunit1 = (trimws(mapunit1)),
         mapunit2 = (trimws(mapunit2))) %>%
  mutate(mapunit12 = paste0(mapunit1,"_", mapunit2)) %>%
  mutate(mapunit12 = gsub("_NA","", mapunit12)) %>%
  mutate(mapunit12 = gsub("_$","", mapunit12)) %>%
  dplyr::select(-TID, -ID)

st_write(processed_transects,  file.path(output_pnts_dir, "proc_s1_transects_cont.gpkg"), 
         delete_layer = TRUE)

st_write(points_final, file.path(output_pnts_dir, "s1_pts_cont.gpkg"), delete_layer = TRUE)

```

## create a summary of calls

The following report summarises field calls by length (transects). A summary of raw point calls, on which transects were collected is include in the Appendices. 

```{r import base files to run analysis, tidy = TRUE, warning=FALSE, include = FALSE}
## 1) Import consolidated data and linear transects

s1_trans_QA <- st_read(list.files(output_pnts_dir, pattern = "transects_QA.gpkg$", full.names = TRUE, recursive = FALSE))

s1_trans_cont <- st_read(list.files(output_pnts_dir, pattern = "transects_cont.gpkg$", full.names = TRUE, recursive = FALSE))

s1_trans_QA <- st_read(s1_trans_QA, quiet = TRUE) %>% mutate(data_type ="QA")
s1_trans_cont  <- st_read(s1_trans_cont, quiet = TRUE)%>% mutate(data_type ="contractor")

# convert the vector data to raster for sampling. 

if("ID" %in% colnames(s1_trans_QA) == FALSE){
        print ("adding ID column")
        s1_trans_QA <- s1_trans_QA %>% 
          mutate(ID = seq(1, length(s1_trans_QA$order), 1))
      }
       
      s1_trans_QA_id <- st_drop_geometry(s1_trans_QA)
      
      lBuff <- s1_trans_QA %>% 
          sf::st_buffer(., dist = 2.5, endCapStyle = "FLAT", joinStyle = "MITRE") %>%
          sf::st_cast(.,"MULTIPOLYGON")
      
      rastAll <- fasterize(lBuff, raster_template, field = "ID")
      
      raster_points_xy <- as.data.frame(rasterToPoints(rastAll)) %>% 
        st_as_sf(coords = c("x", "y"), crs = 3005) %>%
        merge( s1_trans_QA_id, by.x = names(rastAll), by.y = "ID") 
      
      # format transect id names 
      raster_points_xy <- format_transect(raster_points_xy) 
      
      st_write(raster_points_xy, file.path(output_pnts_dir, "QA_transect_all_pts.gpkg"), 
                              delete_layer = TRUE)
      
      
      
## Repeat for Contractor data
   
         
if("ID" %in% colnames(s1_trans_cont) == FALSE){
        print ("adding ID column")
        s1_trans_cont <- s1_trans_cont %>% 
          mutate(ID = seq(1, length(s1_trans_cont$order), 1))
        }
       
      s1_trans_cont_id <- st_drop_geometry(s1_trans_cont )
      
      lBuff <- s1_trans_cont %>% 
          sf::st_buffer(., dist = 2.5, endCapStyle = "FLAT", joinStyle = "MITRE") %>%
          sf::st_cast(.,"MULTIPOLYGON")
      
      rastAll <- fasterize(lBuff, raster_template, field = "ID")
      
      raster_points_xy <- as.data.frame(rasterToPoints(rastAll)) %>% 
        st_as_sf(coords = c("x", "y"), crs = 3005) %>%
        merge(s1_trans_cont_id, by.x = names(rastAll), by.y = "ID") 
      
      # format transect id names 
      raster_points_xy <- format_transect(raster_points_xy) 
      
      st_write(raster_points_xy, file.path(output_pnts_dir, "cont_transect_all_pts.gpkg"), 
                              delete_layer = TRUE)

```
 
 
# Calculate statistics 
 
```{r, echo = FALSE, fig.width = 10, fig.height=10}
## Summary 
qa_pts <- st_read(file.path(output_pnts_dir, "QA_transect_all_pts.gpkg"))
co_pts <- st_read(file.path(output_pnts_dir, "cont_transect_all_pts.gpkg"))

sites_qa <- unique(qa_pts$transect_id)
sites_cont <- unique(co_pts$transect_id)

sites_all <- sites_cont[sites_cont %in% sites_qa]

# manual temp adjustment for testing 
sites_all <- c("ESSFdk 1_3.3_13_cLHS" , "ESSFdk 1_3.3_13_W","ESSFdk 1_4.3_18_cLHS")


# Filter for the sites where both QA and contractors completed


qa_pts <- qa_pts %>% filter(transect_id %in% sites_all) %>% cbind(st_coordinates(.))

co_pts <- co_pts %>% filter(transect_id %in% sites_all) %>% cbind(st_coordinates(.))

allpts <- st_join(qa_pts, co_pts, by = c(X, Y))

# set up table with mapunit1_QA and mapunit1_co (only for overlapping ground)

subpts <- allpts %>%
  dplyr::select(mapunit1.x, mapunit2.x,transition.x, observer.x, comments.x ,
                mapunit1.y,  mapunit2.y, transition.y, observer.y,comments.y ) %>%
  filter(!is.na(mapunit1.y)) %>% # only keep values where there are calls on the same location 
  st_drop_geometry()


  test.pred <-  subpts %>% 
    dplyr::select(mapunit1.x, mapunit2.x, mapunit1.y, mapunit2.y) 
 
  # run accuracy function
  test.acc <- acc_metrix_QA(test.pred)
  
  test.acc
  acc<- test.acc
  
  
  # target = unit
  # trans.tot = no of points made by QA
  # pred.tot = no. of points made by contractor
  # trans.sum = no of points of the total transect (where both QA and Contractor made a call)
  # aspat_p 
       
 [5] "aspat_p"           "aspat_p_overall"   "aspat_p_unit_pos"  "aspat_p_meanacc"  
 [9] "spat_p_overall"    "spat_p_correct"    "spat_p_unit_pos"   "spat_p_meanacc"   
[13] "spat_pa_correct"   "spat_pa_total"     "spat_pa_overall"   "spat_pa_unit_pos" 
[17] "spat_pa_meanacc"   "aspat_pa_extra"    "aspat_pa_total"    "aspat_pa_pred"    
[21] "aspat_pa_unit_pos" "aspat_pa_overall"  "aspat_pa_meanacc"  "accuracy" 



  
  

trans_f<- all_trans %>%
  mutate(mapunit2 = ifelse(mapunit2 == "", NA, mapunit2),
         mapunit1 = ifelse(mapunit1 == "", NA, mapunit1)) %>%
  group_by(mapunit1, mapunit2) %>%
  dplyr::summarise(length = sum(length)) %>%
  ungroup()
  
# primary calls: 
trans_f_p <- trans_f %>%
  filter(is.na(mapunit2)) 

# format secondary calls:
trans_f_s12 <- trans_f %>%
  filter(!is.na(mapunit2)) %>%
  mutate(order = "mapunit12")
       
trans_f_s21 <- trans_f %>%
  filter(!is.na(mapunit2)) %>%
  mutate(order = "mapunit21",
         mapunit3 = mapunit1, 
         mapunit4 = mapunit2) %>%
  dplyr::select(-c(mapunit1, mapunit2)) %>%
  dplyr::rename(mapunit1 = mapunit4, 
                mapunit2 = mapunit3)

trans_secondary <- bind_rows(trans_f_s12, trans_f_s21) %>%
  arrange(mapunit1, mapunit2) %>%
  group_by(mapunit1, mapunit2) %>%
  dplyr::summarise(length = sum(length)) 

trans_secondary_f1 <- trans_secondary %>%
  group_by(length) %>%
  arrange(length, group_by = TRUE) %>%
  aggregate(.~length, . , FUN=head, 1) %>%
  dplyr::select(mapunit1, mapunit2, length)%>%
  arrange(mapunit1) %>%
  mutate(length = round(length, 2))


trans_summary <- bind_rows(trans_f_p, trans_secondary_f1) 
  
trans_total <- trans_summary %>%
  arrange(mapunit1) %>%
  rowwise() %>%
  mutate(bgc = unlist(strsplit(mapunit1, "_"))[1]) %>%
  mutate(bgc = ifelse(str_detect(mapunit1, "_"), bgc, "non-forest")) %>%
  ungroup() %>%
  mutate(length = as.numeric(round(length, 1))) %>%
  filter(!is.na(mapunit1)) %>%
  mutate(mapunit12 = ifelse(is.na(mapunit2), mapunit1, paste0(mapunit1, "/", mapunit2)))

bgc_list <- unique(trans_total$bgc)

```


```{r}

    
## Accuracy measures

The overall map accuracy was calculated but determining the percent correct for each map unit by comparing transect data (held-out slice).

### Types of accuracy 
Several types of accuracy measures were calculated;

1) unweighted (aspatial): this is equivalent to traditional AA where proportion of map units are compared. Aspatial_acc is the accuracy per slice (i.e total accuracy over the site). Aspatial_meanacc is the accuracy based on the average of map units (ie: 100% correct = 0% correct).

2) area weighted (spatial) (spat_p): this compares spatial equivalents for the primary call for each pixal/point predicted.

3) spatial primary/alt calls (spat_pa). This assigns a value if the alternate call matches the predicted call. 

4) fuzzy spatially explicit accuracy: we tested an alternate accuracy measure (spat_fp) to account for calls which were similar (on the edatopic position) to the correct calls. In this case scores could be awarded for on a sliding scale from 1 (Correct) to 0 (no where close) with partial credit assigned to closely related mapunits. Note this requires a matrix which specifies the similarity between all combinations of possible calls. This was also calculated for primary and alternate calls (spat_fpa)


## Comparison of CV and test metrics

```{r, cv vs test overall accuracy with confidence intervals, echo = FALSE, eval = TRUE}
#Compare the CV metrics to test metrics to see model fit 

acc_sum <- acc %>%
    dplyr::mutate(across(ends_with("overall"), ~.x *100)) %>%
    dplyr::mutate(across(ends_with("meanacc"), ~.x *100)) %>%
    dplyr::select(slice, acc_type, transect_no,
                  aspat_p_overall,  aspat_p_meanacc, 
                  aspat_fp_overall,  aspat_fp_meanacc,
                  spat_p_overall, spat_p_meanacc,
                  spat_pf_overall,  spat_pf_meanacc, 
                  aspat_pa_overall,  aspat_pa_meanacc,
                  aspat_fpa_overall, aspat_fpa_meanacc,
                  spat_pa_overall,  spat_pa_meanacc,
                  spat_fpa_overall, spat_fpa_meanacc ) %>%
  distinct() 
  
acc_sum_long <- acc_sum %>%
    pivot_longer(cols = where(is.numeric), names_to = "accuracy_type", values_to = "value") %>%
   filter(!accuracy_type == "transect_no") %>%
  mutate(type = case_when(
    str_detect(accuracy_type, "aspat") ~ "aspatial",
    str_detect(accuracy_type, "spat") ~ "spatial"))  %>%
  mutate(type_model = case_when(
    str_detect(accuracy_type, "_overall") ~ "area-weighted",
    str_detect(accuracy_type, "_meanacc") ~ "unweighted")) %>%
  mutate(accuracy_type_label = case_when(
    str_detect(accuracy_type, "_p_") ~ "p",
    str_detect(accuracy_type, "_pa_") ~ "pa",
    str_detect(accuracy_type, "_fp_") ~ "fp",
    str_detect(accuracy_type, "_pf_") ~ "fp",
    str_detect(accuracy_type, "_fpa_") ~ "fpa")) %>%
  mutate(type_label = paste0(type, "_", type_model))

# calculate the weighted mean and st dev summary 

acc_wt_ave <- acc_sum %>%
  group_by(acc_type) %>%
  dplyr::summarise(dplyr::mutate(across(where(is.numeric), ~ weighted.mean(.x, transect_no, na.rm = FALSE)))) %>%
  pivot_longer(cols = where(is.numeric), names_to = "accuracy_type", values_to = "ave_wt") 
            
acc_wt_sd <- acc_sum %>%         
   group_by(acc_type) %>%         
  dplyr::summarise(mutate(across(where(is.numeric), ~ sqrt(wtd.var(.x, transect_no, na.rm = FALSE))))) %>%
  pivot_longer(cols = where(is.numeric), names_to = "accuracy_type", values_to = "sd_wt") 
          
acc_wt_sum <- left_join(acc_wt_ave, acc_wt_sd ) %>%
  filter(!accuracy_type == "transect_no") 

acc_wt_sum <- acc_wt_sum  %>%
  mutate(type = case_when(
    str_detect(accuracy_type, "aspat") ~ "aspatial",
    str_detect(accuracy_type, "spat") ~ "spatial")) %>%
  mutate(type_model = case_when(
    str_detect(accuracy_type, "_overall") ~ "area-weighted",
    str_detect(accuracy_type, "_meanacc") ~ "unweighted")) %>%
  mutate(accuracy_type_label = case_when(
    str_detect(accuracy_type, "_p_") ~ "p",
    str_detect(accuracy_type, "_pa_") ~ "pa",
    str_detect(accuracy_type, "_fp_") ~ "fp",
    str_detect(accuracy_type, "_pf_") ~ "fp",
    str_detect(accuracy_type, "_fpa_") ~ "fpa")) %>%
  mutate(type_label = paste0(type, "_", type_model))


# set up order for plots 
acc_sum_long$type_f = factor(acc_sum_long$type_label, levels = c("spatial_area-weighted" ,"aspatial_area-weighted", "spatial_unweighted",  "aspatial_unweighted"))

acc_sum_long$accuracy_type_label = factor(acc_sum_long$accuracy_type_label, levels = c("p","pa", "fp","fpa"))


# plot both the cv and test metrics
p2 <- ggplot(aes(y = value, x = accuracy_type_label , fill = acc_type), data = acc_sum_long ) + 
   geom_boxplot() +
   facet_wrap(~type_f, scales = "free_x", nrow = 2) +
   geom_hline(yintercept = 65,linetype ="dashed", color = "black") + 
   ggtitle("Accuracy measures (median + quartiles)") + 
   xlab("Mapunit") + ylab("Accuracy") + 
   ylim(-0.05, 100)+
   theme_pem_facet()+ 
   scale_fill_discrete_sequential(palette = "Light Grays")

p2

```

## Aspatial and Spatial (overall and average accuracy types)

```{r, overall accuracy with confidence intervals, echo = FALSE, eval = TRUE}
# plot only the test metrics 
test_sum_long <- acc_sum_long %>%
  dplyr::filter(acc_type == "test_estimate") %>%
  left_join( acc_wt_sum )

test_sum_long$accuracy_type_label = factor(test_sum_long$accuracy_type_label, levels = c("p","pa", "fp","fpa"))


p3 <- ggplot(aes(y = value, x = accuracy_type_label), data = test_sum_long) + 
   geom_boxplot() +
   facet_wrap(~type_f, scales = "free_y", nrow = 2) +
   geom_hline(yintercept = 65,linetype ="dashed", color = "black") + 
   ggtitle("Accuracy measures (median + quartiles)") + 
   xlab("Mapunit") + ylab("Accuracy") + 
   ylim(-0.05, 100) +
   #theme_bw() + 
   theme_pem_facet() + 
   scale_fill_discrete_sequential(palette = "Light Grays")+ 
   geom_point(aes(y = ave_wt, x = accuracy_type_label), data = test_sum_long, shape = 5, size = 2) 
  

p3

```


## Accuracy per mapunit

We can compare map unit accuracy levels to assess under or acceptable performance per map units. 

```{r generate overall mapunit, echo = FALSE, eval = TRUE}

# map unit plots: 

mu_acc <- acc %>%
    dplyr::select(slice, target, acc_type, transect_no,
                 aspat_p_unit_pos, aspat_p_meanacc, 
                 aspat_fp_unit_pos, aspat_fp_meanacc,
                 aspat_pa_unit_pos, aspat_pa_meanacc,
                 aspat_fpa_unit_pos,aspat_fpa_meanacc,
                 spat_p_unit_pos, spat_p_meanacc,
                 spat_pf_unit_pos, spat_pf_meanacc, 
                 spat_pa_unit_pos, spat_pa_meanacc,
                 spat_fpa_unit_pos, spat_fpa_meanacc
                  ) %>%
  dplyr::filter(acc_type == "test_estimate") %>%
  pivot_longer(cols = where(is.numeric), names_to = "accuracy_type", values_to = "value") %>% filter(accuracy_type != "transect_no")


mu_acc <- mu_acc %>%
  mutate(type = case_when(
    str_detect(accuracy_type, "aspat") ~ "aspatial",
    str_detect(accuracy_type, "spat") ~ "spatial")) %>%
  mutate(type_model = case_when(
    str_detect(accuracy_type, "_unit_pos") ~ "mapunit",
    str_detect(accuracy_type, "_meanacc") ~ "unweighted")) %>%
  mutate(accuracy_type_label = case_when(
    str_detect(accuracy_type, "_p_") ~ "p",
    str_detect(accuracy_type, "_pa_") ~ "pa",
    str_detect(accuracy_type, "_fp_") ~ "fp",
    str_detect(accuracy_type, "_pf_") ~ "fp",
    str_detect(accuracy_type, "_fpa_") ~ "fpa")) %>%
  mutate(type_label = paste0(type, "_", type_model))
 

mu_unit <- mu_acc %>%
  filter(type_model == "mapunit") %>%
  dplyr::select(-c(acc_type, type_model ))

## set up order for plots 
#bsRes_temp$type_f = factor(bsRes_temp$type_label, levels = #c("spatial_overall","aspatial_overall", "spatial_average", "aspatial_average"))

mu_unit$accuracy_type_label = factor(mu_unit$accuracy_type_label, levels = c("p","pa", "fp","fpa"))


p4 <- ggplot(aes(y = value, x = accuracy_type_label , fill = type), data = mu_unit  ) + 
   geom_boxplot() +
   facet_wrap(~target, scales = "free_x", nrow = 2) +
   ggtitle("Mapunit accuracy measures ") + 
   xlab("accuracy measure") + ylab("Proportion of Accurate calls") + 
   ylim(-0.05, 1)+
   theme_pem_facet()+ 
   scale_fill_discrete_sequential(palette = "Light Grays")
  

p4


```
```


The length of calls per bgc varied, along with proportion of primary and secondary calls. Figures below show the length of calls for all data types grouped by predominant bgc.


```{r, echo = FALSE, fig.width= 12, fig.height = 16}
# plot the freqency 

# points collected per type (primary and secondary calls) all calls
#p <- ggplot(trans_total, aes(y = mapunit12, x = length)) + 
#  facet_wrap(~bgc, scales = "free_x") +    
#  geom_bar(stat = "identity")
#         
#print(p)



for(b in bgc_list) {

 # b = unique(pt_total$bgc)[3]
  pt_total_bgc <- trans_total %>%
    filter(bgc == b)
  
  p <- ggplot(pt_total_bgc , aes(y = reorder(mapunit12, length), x = length)) +
  geom_bar(stat = "identity") + 
  labs(title = b, y = "mapunit")
  
  print(p)
  
#  ggplot(pt_total_bgc , aes(y = mapunit1, x = freq)) +
#  geom_bar(stat = "identity")
     
  #ggsave(file.path(s1_dir, paste0(b, "_point_freq_s1.png")))
  
  
}

```

### Proportion of primary and secondary calls per observer.

```{r echo = FALSE}

# group by observer

all_trans_ob <- all_trans %>%
  group_by(transect_id) %>%
  fill( observer, .direction = "downup")

#all_trans_ob_sum <- all_trans_ob %>%
#   group_by(transect_id, observer) %>%
#    dplyr::summarise(count = n())

trans_ob_p <- all_trans_ob %>%
 # dplyr::filter(is.na(mapunit2)) %>%
  group_by(observer) %>%
  summarise(all.calls = sum(length))

trans_ob_s <- all_trans_ob %>%
  dplyr::filter(!is.na(mapunit2)) %>%
  group_by(observer) %>%
  summarise(sec.calls = sum(length))


trans_obs_type <- left_join(trans_ob_p, trans_ob_s) %>%
  mutate(pc.sec = round((sec.calls/all.calls * 100),1))

trans_obs_type<- trans_obs_type %>%
  mutate(observer = case_when(
    observer == "EAC GP" ~ "EAC",
    observer == "EAC GCP" ~ "EAC",
    observer == "EC" ~ "EAC",
    observer == "CC LJ" ~ "CC",
    observer == "DF LJ" ~ "DF",
    observer =="DJM DF" ~ "DJM",
    observer == "PD AR" ~ "PD", 
    observer =="SRan" ~ "SR",
    observer == "SR/MZ" ~ "SR",
    observer == "WM" ~ "WHM",
        TRUE ~ observer)) %>%
      group_by(observer) %>%
      summarise(across(where(is.numeric), sum)) %>%
  mutate(pc.sec = round(sec.calls/all.calls * 100,1))

```


There were `r length(trans_ob_s$observer)` observers recording data, including unnamed recorders (NA). The table below shows the total calls and the proportion of calls which also had a secondary call. The average proportion of transect length that has a secondary call was `r mean(trans_ob_s$pc.sec)`. Note this data does not include repeat transects, in which multiple observers recorded at the same location. 

```{r, results  = "asis", hide = TRUE, echo = FALSE}

kable(trans_obs_type)

```




```{r, include = FALSE}
# 
# Appendices
# ### Summary of point data 
# 
# Point data represents the raw data collected by observers during transect traverses. This is included for comparison only. 

```


```{r, include = FALSE}
# combine stage 1 and 2 and summarise calls

s1_pt <- st_read(s1_pts, quiet = T) 
s2_pt <- st_read(s2_pts, quiet = T) 

pts <- bind_rows(s1_pt, s2_pt) %>%
  dplyr::select(-order) %>%
  st_drop_geometry() %>%
  filter(!is.na(mapunit1)) %>%
  mutate(mapunit12 = paste0(mapunit1,"_", mapunit2)) %>%
  mutate(mapunit12 = gsub("_NA","", mapunit12)) %>%
  mutate(mapunit12 = gsub("_$","", mapunit12)) 

# proportion of calls with primary and secondary calls

# total type of points with primary and secondary 
pt_type_p <- pts %>%
  group_by(data_type) %>%
  dplyr::summarise(all.calls = n())

#pt_type_i <- pts %>%
#  filter(data_type == "incidental")

pt_type_s <- pts %>%
  filter(!is.na(mapunit2)) %>%
  group_by(data_type) %>%
  dplyr::summarise(sec.calls = n())

pt_type <- left_join(pt_type_p, pt_type_s) %>%
  mutate(percent.sec = sec.calls/all.calls * 100)

pt_types <- unique(pt_type$data_type)

```

```{r, include = FALSE}

# At `r AOI` study area a total `r sum(pt_type$all.calls)` point calls were collected during traverse (stage 1 and stage 2) transect collection. This include `r length(pt_type$data_type)` data types : `r pt_types`.
# 
# Of these, `r sum(pt_type$sec.calls)` points were also given an secondary call, representing an average percent of `r round(sum(pt_type$sec.calls)/sum(pt_type$all.calls)*100, 1) ` %).

```


```{r, results  = "asis", hide = TRUE, echo = FALSE,  include = FALSE}

kable(pt_type)

```

```{r, echo = FALSE, warnings = FALSE, messages = FALSE, fig.width = 12,fig.height = 16, include = FALSE}
# calculate the frequency of all calls with primary and secondary
#The following figures provide a breakdown of primary and secondary call #frequency of point data collected. Note in the figures below secondary #calls are presented twice in each figure (ie: CX_AM and AM_CX). 

# stage 1 
pt_f <- pts %>%
  filter(data_type == "s1") %>%
  mutate(mapunit2 = ifelse(mapunit2 == "", NA, mapunit2),
         mapunit1 = ifelse(mapunit1 == "", NA, mapunit1)) %>%
  group_by(mapunit1, mapunit2) %>%
  dplyr::summarise(freq = n())
  

# primary calls: 
pt_f_p <- pt_f %>%
  ungroup() %>%
  filter(is.na(mapunit2)) %>%
  dplyr::select(-mapunit2)

# format secondary calls:
pt_f_s <- pt_f %>%
  filter(!is.na(mapunit2)) %>%
  mutate(mapunit12 = paste0(mapunit1, "_", mapunit2),
         mapunit21 = paste0(mapunit2, "_", mapunit1)) %>%
  ungroup()

# grab mapunit12
pt_f_s1 <- pt_f_s %>%
  dplyr::select(mapunit12, freq) %>%
  mutate(type = "mapunit12")

# grab mapunit21
pt_f_s2 <- pt_f_s %>%
  dplyr::select(mapunit21, freq)%>%
  mutate(type = "mapunit21",
         mapunit12 = mapunit21 ) %>%
 dplyr::select(-c(mapunit21))

#duplicates <- c(pt_f_s2$mapunit12)

pt_secondary <- bind_rows(pt_f_s1, pt_f_s2) %>%
  arrange(mapunit12) %>%
  group_by(mapunit12) %>%
  dplyr::summarise(freq = sum(freq)) %>%
  dplyr::rename(mapunit1 = mapunit12)

#pt_secondary_tidy <- pt_secondary %>%
#  mutate(map1 = str_split(mapunit1, "[:digit:]\\_", n = 2))
#pt_secondary_tidy$map1


#note these appear twice for each combination! Need to remove these!

pt_total <- bind_rows(pt_f_p, pt_secondary) %>%
  arrange(mapunit1) %>%
mutate(bgc = case_when(
    str_detect(mapunit1,"ESSFmc_") ~ "ESSFmc",
    str_detect(mapunit1, "SBS") ~  "SBS",
    str_detect(mapunit1,"ESSFdc2") ~ "ESSFdc",
    str_detect(mapunit1,"ESSFmh") ~ "ESSFmh",
    str_detect(mapunit1,"ICHmk1") ~ "ICHmk",
    str_detect(mapunit1,"IDFdm1") ~ "IDFdm",
    str_detect(mapunit1,"MSdm1") ~ "MSdm",
    TRUE ~ "non-forest"
    )) %>%
  dplyr::filter(!is.na(mapunit1))
  
# points collected per type (primary and secondary calls) all calls
#p <- ggplot(pt_total, aes(y = mapunit1, x = freq)) + 
#  facet_wrap(~bgc, scales = "free_x") +    
#  geom_bar(stat = "identity")
#         
#p

# points collected per bgc (primary and secondary calls) 
bgc_list <- unique(pt_total$bgc)
plot_list <- vector(length = length(bgc_list), mode = "list")
names(plot_list) <- bgc_list


for(b in bgc_list) {

 # b = unique(pt_total$bgc)[3]
  pt_total_bgc <- pt_total %>%
    filter(bgc == b) %>%
    arrange(freq)
  
  p <- ggplot(pt_total_bgc , aes(y = reorder(mapunit1, freq), x = freq)) +
  geom_bar(stat = "identity") + 
  labs(title = b, y = "mapunit")
  
  print(p)
  
#  ggplot(pt_total_bgc , aes(y = mapunit1, x = freq)) +
#  geom_bar(stat = "identity")
     
  #ggsave(file.path(s1_dir, paste0(b, "_point_freq_s1.png")))
  
  
}

```
